Move functions into a helper module
